used jwt as authentication during deletion of post

+++ controllers/APIs/v1/posts_api.js
@@ -24,34 +24,26 @@ module.exports.destroy =async function(req, res) {
     try {
         let post = await Post.findById(req.params.id);
         
-        //if(post.user == req.user.id){
+        if(post.user == req.user.id){
             post.remove();
             //return res.redirect('back');
             await Comment.deleteMany({post: req.params.id}) ;
 
 
 
-            // if(req.xhr){
-            //     return res.status(200).json({
-            //         data: {
-            //             post_id: req.params.id
-            //         },
-            //         message: "post deleted!"
-            //     });
-            // }
 
 
 
-            //req.flash('success', 'Post and associated comment deleted');
+            
 
                 return res.json('200',{
                     message: "post and associated comment deleted"
                 });
-        // } else {
-        //     req.flash('error', 'You cannot delete this post');
-
-        //     return res.redirect('back');
-        // }
+        } else {
+            return res.json(401,{
+                message: "you cannot delete thsi post"
+            })
+        }
     } catch(err) {
         //req.flash('error', err);
